moved new instance logic to wasmer_instance.rs
